+++ mud-contracts/world/mud.config.ts
@@ -474,8 +474,10 @@ export default defineWorld({
     KillMailTable: {
       schema: {
         killMailId: "uint256",
-        killer: "address",
-        victim: "address",
+      },
+      valueSchema: {
+        killerCharacterId: "uint256",
+        victimCharacterId: "uint256",
         lossType: "KillMailLossType",
         solarSystemId: "uint256",
         killTimestamp: "uint256",