+++ mud-contracts/world/test/kill-mail/KillMail.t.sol
@@ -40,8 +40,8 @@ contract KillMailTest is MudTest {
 
   function testRegisterKill(
     uint256 killMailId,
-    address killer,
-    address victim,
+    uint256 killerCharacterId,
+    uint256 victimCharacterId,
     uint256 solarSystemId,
     uint256 killTimestamp
   ) public {
@@ -50,8 +50,8 @@ contract KillMailTest is MudTest {
     KillMailLossType lossType = KillMailLossType.SHIP;
 
     KillMailTableData memory killMailTableData = KillMailTableData({
-      killer: killer,
-      victim: victim,
+      killerCharacterId: killerCharacterId,
+      victimCharacterId: victimCharacterId,
       lossType: lossType,
       solarSystemId: solarSystemId,
       killTimestamp: killTimestamp
@@ -61,8 +61,8 @@ contract KillMailTest is MudTest {
 
     KillMailTableData memory storedKillMailTableData = KillMailTable.get(killMailId);
 
-    assertEq(killMailTableData.killer, storedKillMailTableData.killer);
-    assertEq(killMailTableData.victim, storedKillMailTableData.victim);
+    assertEq(killMailTableData.killerCharacterId, storedKillMailTableData.killerCharacterId);
+    assertEq(killMailTableData.victimCharacterId, storedKillMailTableData.victimCharacterId);
     assertEq(killMailTableData.killTimestamp, storedKillMailTableData.killTimestamp);
   }